fixtures moved from tests.py to conftest.py
